# API/swagger/spec.json
{"openapi":"3.0.0","info":{"title":"api","version":"1.0.0"},"paths":{"/book":{"get":{"tags":["Book"],"security":[{"bearerAuth":[]}],"parameters":{"schema":{"type":"string"}},"responses":{"200":{"$ref":"#/components/responses/BooksFound"},"400":{"$ref":"#/components/responses/ErrorJWT"},"401":{"$ref":"#/components/responses/MissingJWT"},"404":{"description":"Aucun livre trouvé"},"500":{"description":"Erreur serveur"}}},"post":{"tags":["Livre"],"security":[{"bearerAuth":[]}],"requestBody":{"$ref":"#/components/requestBodies/BookToUpdate"},"responses":{"201":{"$ref":"#/components/responses/BookAdded"},"204":{"$ref":"#/components/responses/BookUpdated"},"400":{"description":"- Fichier image non valide\n- ISBN manquant\n- Formulaire non valide\n- Le JWT n'est pas valide\n"},"401":{"$ref":"#/components/responses/MissingJWT"},"403":{"$ref":"#/components/responses/mustBeAdmin"},"500":{"description":"Erreur serveur"}}},"delete":{"tags":["Livre"],"security":[{"bearerAuth":[]}],"responses":{"204":{"$ref":"#/components/responses/BookDeleted"},"400":{"$ref":"#/components/responses/ErrorJWT"},"401":{"$ref":"#/components/responses/MissingJWT"},"403":{"$ref":"#/components/responses/mustBeManager"},"500":{"description":"Erreur serveur"}}}},"/book/{id}":{"get":{"tags":["Livre"],"security":[{"bearerAuth":[]}],"parameters":[{"name":"id","description":"ISBN d'un livre","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"$ref":"#/components/responses/BookFound"},"400":{"$ref":"#/components/responses/ErrorJWT"},"401":{"$ref":"#/components/responses/MissingJWT"},"404":{"description":"Livre non trouvé"},"500":{"description":"Erreur serveur"}}}},"/reviews/{id}":{"get":{"tags":["Reviews"],"security":[{"bearerAuth":[]}],"parameters":[{"in":"path","name":"id","schema":{"type":"integer"},"required":true,"description":"The review id"}],"responses":{"200":{"$ref":"#/components/responses/ReviewFound"},"404":{"description":"The review was not found"},"500":{"description":"Some server error"}}},"patch":{"tags":["Reviews"],"parameters":[{"name":"id","in":"path","description":"ID of the review to update","required":true,"schema":{"type":"integer"}}],"security":[{"bearerAuth":[]}],"requestBody":{"$ref":"#/components/requestBodies/ReviewAUpdate"},"responses":{"204":{"$ref":"#/components/responses/ReviewUpdated"},"400":{"$ref":"#/components/responses/ErrorJWT"},"401":{"$ref":"#/components/responses/MissingJWT"},"403":{"$ref":"#/components/responses/mustBeAdmin"},"404":{"description":"Review not found"},"500":{"description":"Internal error"}}},"delete":{"tags":["Reviews"],"security":[{"bearerAuth":[]}],"parameters":[{"in":"path","name":"id","schema":{"type":"integer"},"required":true,"description":"The review id"}],"responses":{"204":{"$ref":"#/components/responses/ReviewDeleted"},"400":{"$ref":"#/components/responses/ErrorJWT"},"401":{"$ref":"#/components/responses/MissingJWT"},"403":{"$ref":"#/components/responses/mustBeAdmin"},"404":{"description":"The review was not found"},"500":{"description":"Some error happened"}}}},"/reviews":{"get":{"tags":["Reviews"],"security":[{"bearerAuth":[]}],"responses":{"200":{"$ref":"#/components/responses/allReviewsFound"},"500":{"description":"Some server error"}}},"post":{"tags":["Reviews"],"security":[{"bearerAuth":[]}],"requestBody":{"$ref":"#/components/requestBodies/ReviewAAjoute"},"responses":{"201":{"$ref":"#/components/responses/ReviewAjoute"},"400":{"$ref":"#/components/responses/ErrorJWT"},"401":{"$ref":"#/components/responses/MissingJWT"},"403":{"$ref":"#/components/responses/mustBeAdmin"},"500":{"description":"Internal error"}}}}},"components":{"responses":{"mustBeAdmin":{"description":"Access denied to all non-admin users"},"ErrorJWT":{"description":"le JWT n'est pas valide"},"MissingJWT":{"description":"le JWT n'est pas présent"}},"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}},"tags":[]}